azure-concept-summary/requirements.txt
openai
langchain
langchain-openai
python-dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")